pyproject.toml
[project]
name = "aplicationtracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth[with-social]>=7.0.1",
    "django>=5.2.4",
    "django-allauth==64.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg-binary>=3.2.9",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
]
